feat: add init web output

+++ FindPortProcesses.ps1
@@ -1,10 +1,15 @@
 param([switch]$kill)
-$ports = @(3000, 5099, 5000, 11119, 5173, 8080)
+$ports = @(3000, 5099, 5000, 11119, 5173, 8080, 1111)
 
 foreach ($port in $ports) {
     $connection = Get-NetTCPConnection -LocalPort $port -ErrorAction SilentlyContinue
     if($connection) {
-        $process = Get-Process -Id ($connection).OwningProcess
+        $process = Get-Process -Id ($connection)
+        if(!$process) {
+            continue
+        }
+        
+        $process = $process.OwningProcess
         if ($process -and $kill) {
             Write-Host "Killing process $($process.Id) running on port $port"
             Stop-Process -Id $process.Id -Force
